Move iconlbi and iconPaths into symbology folder
