+++ src/symbology/iconPaths.ts
